requirements.txt
python-dotenv==1.0.0
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.3
openai==1.3.7
chromadb==0.4.22
beautifulsoup4==4.12.2
requests==2.31.0
python-telegram-bot==20.7
youtube-transcript-api==0.6.1
tiktoken==0.5.2 